+++ site-cookbooks/rvm/definitions/rvm_ruby.rb
@@ -1,0 +1,6 @@
+define :rvm_ruby, :version => nil do
+  version = params[:version] || params[:name]
+  execute "rvm install #{version}" do
+    environment 'PATH' => node[:rvm][:path]
+  end
+end