+++ site-cookbooks/sphinx/recipes/default.rb
@@ -1,12 +1,11 @@
 #
 # Cookbook Name:: sphinx
-# Recipe:: install
+# Recipe:: default
 #
 # Copyright 2017, Leonis0813
 #
 # All rights reserved - Do Not Redistribute
 #
-
 pyenv = node[:sphinx][:pyenv]
 python = node[:sphinx][:python]
 
@@ -16,12 +15,11 @@ git pyenv[:root] do
   not_if { File.exists?(pyenv[:root]) }
 end
 
-unless File.exists?("#{pyenv[:root]}/versions/#{python[:version]}")
-  ['init -', "install #{python[:version]}"].each do |command|
-    execute "pyenv #{command}" do
-      environment 'PYENV_ROOT' => pyenv[:root],
-                  'PATH' => "#{pyenv[:root]}/bin:/usr/bin:/bin"
-    end
+['init -', "install #{python[:version]}"].each do |command|
+  execute "pyenv #{command}" do
+    environment 'PYENV_ROOT' => pyenv[:root],
+                'PATH' => "#{pyenv[:root]}/bin:/usr/bin:/bin"
+    not_if { File.exists?("#{pyenv[:root]}/versions/#{python[:version]}") }
   end
 end
 
@@ -31,5 +29,6 @@ node[:sphinx][:packages].each do |package|
   execute "pyenv global #{python[:version]} && pip install #{package}" do
     environment 'PYENV_ROOT' => pyenv[:root],
                 'PATH' => "#{pyenv[:root]}/versions/#{python[:version]}/bin:#{pyenv[:root]}/bin:/usr/bin:/bin"
+    not_if "pyenv global #{python[:version]} && pip list | cut -d ' ' -f 1 | grep -i #{package.split('==').first}"
   end
 end